+++ src/main/resources/db/migration/V07__nye_felter.sql
@@ -3,7 +3,7 @@ ALTER TABLE avtale ADD COLUMN antall_dager_per_uke DOUBLE PRECISION;
 
 -- En nyanse vi ikke har i avtaleløsningen er at stillingprosent-feltet gjelder mentor for mentoravtaler,
 -- ikke deltakeren! Det kan vi jo rette opp i her.
-ALTER TABLE avtale ADD COLUMN stillingprosent INTEGER;
+ALTER TABLE avtale ADD COLUMN stillingprosent DOUBLE PRECISION;
 
 -- [jooq ignore start]
 UPDATE avtale SET antall_dager_per_uke = (raw_json->>'antallDagerPerUke')::NUMERIC
